refactor: Fix Checkstyle errors
