+++ src/main/java/chatbot/task/Event.java
@@ -1,11 +1,11 @@
 package chatbot.task;
 
-import chatbot.exception.ChatBotException;
-
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
+import chatbot.exception.ChatBotException;
 
 /**
  * Represents a task that spans a period of time.
