+++ src/test/java/chatbot/task/EventTest.java
@@ -1,10 +1,9 @@
 package chatbot.task;
 
-import org.junit.jupiter.api.Test;
-
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
+
+import org.junit.jupiter.api.Test;
 
 public class EventTest {
     @Test
